// app/api/openai/types.ts
import { MultimodalContent } from "../api";
import { ModelSize, DalleQuality, DalleStyle } from "@/app/typing";

export interface OpenAIListModelResponse {
  object: string;
  data: Array<{ id: string; object: string; root: string }>;
}

export interface RequestPayload {
  messages: {
    role: "developer" | "system" | "user" | "assistant";
    content: string | MultimodalContent[];
  }[];
  stream?: boolean;
  model: string;
  temperature: number;
  presence_penalty: number;
  frequency_penalty: number;
  top_p: number;
  max_tokens?: number;
  max_completion_tokens?: number;
}

export interface DalleRequestPayload {
  model: string;
  prompt: string;
  response_format: "url" | "b64_json";
  n: number;
  size: ModelSize;
  quality: DalleQuality;
  style: DalleStyle;
}